+++ NEWS.md
@@ -52,6 +52,7 @@ sudo make install
 * Kuviz page:
   * [Dashboard] Add Kuviz page (I) - STRUCTURE ([#15170](https://github.com/CartoDB/cartodb/issues/15170))
   * [Dashboard] Add Kuviz page (II) - LIST ([#15171](https://github.com/CartoDB/cartodb/issues/15171))
+  * [Dashboard] Add Kuviz page (IIIa) - QUICK ACTIONS Delete ([#15172](https://github.com/CartoDB/cartodb/issues/15172))
 
 ### Bug fixes / enhancements
 * Change utm_parameters ([#15146](https://github.com/CartoDB/cartodb/pull/15146))